+++ src/main/java/roomescape/domain/reservation/entity/Reservation.java
@@ -1,6 +1,7 @@
 package roomescape.domain.reservation.entity;
 
 import lombok.*;
+import roomescape.domain.time.entity.Time;
 
 import java.time.LocalDate;
