+++ src/main/java/roomescape/domain/reservation/service/ReservationService.java
@@ -4,15 +4,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import roomescape.domain.reservation.dto.response.GetReservationResponse;
-import roomescape.domain.reservation.dto.response.GetTimesResponse;
 import roomescape.domain.reservation.entity.Reservation;
-import roomescape.domain.reservation.entity.Time;
+import roomescape.domain.time.entity.Time;
 import roomescape.domain.reservation.repository.ReservationRepositoryJdbc;
-import roomescape.domain.reservation.repository.TimesRepositoryJdbc;
+import roomescape.domain.time.repository.TimesRepositoryJdbc;
 import roomescape.global.BusinessException;
 
 import java.time.LocalDate;
-import java.time.LocalTime;
 import java.util.List;
 
 import static roomescape.global.ErrorCode.*;
@@ -56,27 +54,5 @@ public class ReservationService {
         reservationRepositoryJdbc.deleteById(id);
     }
 
-    @Transactional
-    public Time saveTime(final LocalTime time) {
-        return timeRepositoryJdbc.save(time);
-    }
 
-    public List<GetTimesResponse> findAllTimes() {
-        return timeRepositoryJdbc.findAll()
-                .stream()
-                .map(GetTimesResponse::from)
-                .toList();
-    }
-
-    @Transactional
-    public void deleteTimes(final long timesId) {
-        timeRepositoryJdbc.deleteById(timesId);
-    }
-
-
-    public long findTimes(final long timesId) {
-        return timeRepositoryJdbc.findById(timesId)
-                .orElseThrow(() -> new BusinessException(timesId, "timesId", TIME_NOT_FOUND))
-                .getId();
-    }
 }
