+++ src/main/java/roomescape/domain/time/entity/Time.java
@@ -1,4 +1,4 @@
-package roomescape.domain.reservation.entity;
+package roomescape.domain.time.entity;
 
 import lombok.*;
