+++ src/main/java/roomescape/domain/time/repository/TimesRepository.java
@@ -1,6 +1,6 @@
-package roomescape.domain.reservation.repository;
+package roomescape.domain.time.repository;
 
-import roomescape.domain.reservation.entity.Time;
+import roomescape.domain.time.entity.Time;
 
 import java.time.LocalTime;
 import java.util.List;
